@@ -1,64 +1,39 @@
 name: MJN
 description: A new Flutter application.
-
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  splashscreen: ^1.3.5
-  get: 4.3.8
-  get_storage: ^2.0.3
-  http: ^0.13.4
-  floor: ^1.2.0
+  get: 4.7.2
+  get_storage: ^2.1.1
+  http: ^1.5.0
+  floor: ^1.5.0
   analyzer: ^2.2.0
-  flutter_local_notifications: ^5.0.0+4
-  url_launcher: ^6.0.12
+  flutter_local_notifications: ^18.0.1
+  url_launcher: ^6.3.2
   readmore: ^2.1.0
   event_bus: ^2.0.0
-  webview_flutter: ^3.0.0
-  flutter_multi_carousel: ^1.0.0
-  full_screen_image: ^1.0.2
-  carousel_slider: ^4.0.0
+  webview_flutter: ^4.13.0
+  full_screen_image: ^2.0.0
+  carousel_slider: ^5.1.1
   smooth_page_indicator: ^1.0.0+2
   store_redirect: ^2.0.0
 
-  flutter_neumorphic:
-    git:
-      url: "https://github.com/alionour/Flutter-Neumorphic"
-
-
   plugin_platform_interface: ^2.0.0
-#  firebase_core: ^1.3.0
-#  firebase_messaging: ^10.0.0
-  firebase_core: ^1.10.0
-  firebase_messaging: ^11.1.0
-  firebase_core_platform_interface: ^4.0.1
-  firebase_messaging_platform_interface: ^3.0.0
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  firebase_core: ^4.0.0
+  firebase_messaging: ^16.0.0
+  firebase_core_platform_interface: ^6.0.0
+  firebase_messaging_platform_interface: ^4.7.0
   cupertino_icons: ^1.0.2
   lottie: ^1.1.0
-  android_intent_plus: ^3.0.2
-  flutter_email_sender: ^5.0.2
+  android_intent_plus: ^5.3.1
+  flutter_email_sender: ^7.0.0
+  file: ^7.0.1
 
 
 
@@ -139,31 +114,3 @@ flutter:
     - assets/images/account_icon.png
     - assets/images/logout_icon.png
     - assets/images/mjn_app_icon.png
-
-
-
-        # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
