@@ -1,6 +1,3 @@
-// database.dart
-
-// required package imports
 import 'dart:async';
 import 'package:MJN/models/notificationModelVO.dart';
 import 'package:MJN/presistence/dao/NotificationDao.dart';
@@ -10,8 +7,7 @@ import 'package:sqflite/sqflite.dart' as sqflite;
 
 part 'MyDB.g.dart';
 
-@Database
-  (version: 1, entities: [NotificationModelVO])
+@Database(version: 1, entities: [NotificationModelVO])
 abstract class MyDatabase extends FloorDatabase {
-  NotificationDao get notiDao;
+NotificationDao get notiDao;
 }
