@@ -5,7 +5,6 @@ import 'package:MJN/views/SecondLoginView.dart';
 import 'package:MJN/views/SignUpView.dart';
 import 'package:MJN/views/TabView.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 
@@ -114,7 +113,7 @@ class _SecondLoginVIewState extends State<SecondLoginVIew> {
                 width: MediaQuery.of(context).size.width,
                 height: 40,
                 alignment: Alignment.center,
-                child: NeumorphicButton(
+                child: ElevatedButton(
                   onPressed: () {
                     Navigator.of(context).pushReplacementNamed(TabScreens.routeName);
                   },
@@ -125,13 +124,8 @@ class _SecondLoginVIewState extends State<SecondLoginVIew> {
                         fontSize: 14,
                         color: Colors.white),
                   ),
-                  style: NeumorphicStyle(
-                    shape: NeumorphicShape.flat,
-                    boxShape: NeumorphicBoxShape.roundRect(
-                        BorderRadius.circular(14)),
-                    color: Colors.blue,
-                    depth: 8,
-//                lightSource: LightSource.topLeft,
+                  style: ElevatedButton.styleFrom(
+                      backgroundColor: Colors.blue
                   ),
                 ),
               ),
@@ -155,9 +149,11 @@ class _SecondLoginVIewState extends State<SecondLoginVIew> {
               width: 75,
               margin: EdgeInsets.only(bottom: 37, right: 30, top: 33),
               padding: EdgeInsets.all(3),
-              child: Neumorphic(
-                style: NeumorphicStyle(
-                    color: Colors.white, lightSource: LightSource.topLeft),
+              child: ElevatedButton(
+                style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.white
+                ),
+                onPressed: () {  },
                 child: DropdownButtonFormField<String>(
                   isExpanded: true,
                   value: selectedLang,
